Reorder spherebuilder
